[New][Fix] Convolution is fixes and there are new plots for animations
The convolution was being calculated wrong before and now it is being calculated right by cutting the signal before it starts repeating.

There are also new plots for the animations. A plot with every animation used in the frame is saved under /image/stack and a plot with the frame frequency and animations used in every 1s window is  also stored.