[Major news] Many fixes and new plots for available transitions & cost
There was an error where the dataset was identified as Radius -1 and the code was set to see the dataset as Radius 0. This meant that loads of modifications weren't being applied to the files, as well as exclusions that weren't happening when the overwrite feature was set. Also, there is a new plot showing available transitions between animations. This is all in the main scripts, but there are other changes also. The files are now being recorded on the motion-evaluation folder instead of the Unity project root, which generates meta files. There is also a new cost plot, where each transition's cost is shown.